requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.10
django-bootstrap-v5==1.0.7
djangorestframework==3.12.4
importlib-metadata==3.10.1
Pillow==8.4.0
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.6.0